instance_ssh() will login as root user unless directed
If a value isn't provided, attempt to login as root.
This is because default user for centos is root and
ubuntu checks for this and prints helpful message.

Defaulting to $USER results in ssh trying other keys
and user not seeing helpful message from Ubuntu.

Values in checked (in order of preference):

* argument to instance_ssh()
* value of default-user tag on instance
* $AWS_DEFAULT_USER
* root