cf_get() is now an alias to cf_template()
cf_template() returns stack template from AWS
cf_params() returns stack params from AWS

cf_template() used to be a function used internally to
find the template for a command. It was recently renamed
to cf_x_template to less polute the tab completion for
cf_ functions. I wonder if there is a better name for it?