Added note about "simplification"
I've removed arguments and abstractions that meant users
couldn't read the aws query by just typing `type stack-asgs`.

This breaks DRY (Don't Repeat Yourself) in a way that benefits
users who want to be able to dip in and learn more about how
the functions work.

Bash is a special purpose language for running commands. It's not
Ruby or Python and shouldn't be treated like them.