stack-tag() stack-tag-apply() stack-tag-delete()
stack-tag() returns the value of a named stack tag
stack-tag-apply() creates or updates a named stack-tag
stack-tag-delete() removes a stack-tag

** Warning **
Note that when updating stack tags, you need to specify all of the
ones you wish to exist. I believe they are managed stored within
the CloudFormation service, not the tagging service but this has
not been confirmed.

A known problem with making assumptions about tags based on the tags
returned by queries on resources like EC2 Instances is that if the
tagging service fails, you may get a list of instances but not their
tags. If this were the case with CloudFormation Stack Tags then the
stack-tag-apply() and stack-tag-delete() functions could remove
all tags in the case of a failure in tagging service.

I'll continue to attempt to confirm whether this risk exists and
deal with it if my suspicion that stack-tags are not promne to this
problem is disproved.