add example SSM instance
The SSM Agent that is installed by default is a bit old and doesn't allow
connections.  You can updated the agent by using a UserData section in the
CFN template or you can use the Session Manager console in the "Start a session"
section.  Finally, you can use the CLI by running the following command:

    aws ssm send-command --document-name "AWS-UpdateSSMAgent" \
      --targets '[{"Key":"InstanceIds","Values":["i-0b05de82xxxxxx"]}]' \
      --parameters '{"allowDowngrade":["false"]}' \
      --timeout-seconds 3600 --max-concurrency "50" --max-errors "0" --region us-east-1